+++ src/app/(Main)/talk/page.tsx
@@ -1,6 +1,11 @@
 import TalkSection from './components/TalkSection';
 import TitleSection from './components/TitleSection';
 import { Spacing } from '@/components/Spacing';
+import { Metadata } from 'next';
+
+export const metadata: Metadata = {
+  title: '포즈톡',
+};
 
 export default function Talk() {
   return (
