+++ src/app/layout.tsx
@@ -11,7 +11,6 @@ import QueryProvider from '@/provider/QueryProvider';
 import RecoilContextProvider from '@/provider/RecoilContextProvider';
 
 import type { Metadata } from 'next';
-import Head from 'next/head';
 
 const DEFAULT_OG_TITLE = 'PosePicker';
 const DEFAULT_OG_DESC = '다음 포즈 뭐하지? 포즈피커가 포즈 고민을 해결해 드릴게요!';
@@ -56,13 +55,6 @@ export const metadata: Metadata = {
 export default function RootLayout({ children }: { children: React.ReactNode }) {
   return (
     <html lang="ko">
-      <Head>
-        <meta name="naver-site-verification" content="65f3aba9349cce28018ac7a97d4f87ff00709aa3" />
-        <meta
-          name="google-site-verification"
-          content="MB7qV_Oa4G4gR0jHgjtnE6S4g4blocE2mjo7z-z2f6Q"
-        />
-      </Head>
       <body className="flex min-h-[100dvh] w-screen touch-none justify-center bg-slate-100 py-px">
         <div className="w-full max-w-440 overflow-hidden bg-white text-primary">
           <Suspense>
